# Micro optimisation on parsePublicodes