--- a/packages/api/package.json
+++ b/packages/api/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@publicodes/api",
-	"version": "1.0.0-beta.73",
+	"version": "1.0.0-beta.74",
 	"description": "Publicodes API",
 	"type": "module",
 	"main": "./dist/index.cjs",