--- a/packages/core/bench/index.ts
+++ b/packages/core/bench/index.ts
@@ -1,0 +1,11 @@
+import Publicodes from '../source/index'
+import { bench, group, run } from 'mitata'
+import modeleSocial from 'modele-social'
+
+group('Parsing initial des règles', () => {
+	bench('Modele-social', () => {
+		new Publicodes(modeleSocial)
+	})
+})
+
+run()