--- a/packages/core/package.json
+++ b/packages/core/package.json
@@ -32,7 +32,9 @@
 		"chai": "^4.2.0",
 		"dedent-js": "1.0.1",
 		"intl": "^1.2.5",
+		"mitata": "^0.1.6",
 		"mocha": "^9.2.0",
+		"modele-social": "^1.7.2",
 		"moo": "^0.5.1",
 		"nearley": "^2.19.2",
 		"rimraf": "^3.0.2",
@@ -40,7 +42,7 @@
 		"sinon-chai": "^3.7.0",
 		"ts-node": "^10.9.1",
 		"tsup": "^6.5.0",
-		"typescript": "^4.9.3"
+		"typescript": "^5.2.2"
 	},
 	"scripts": {
 		"build:grammar": "nearleyc ./source/grammar.ne -o ./source/grammar.js",
@@ -50,6 +52,7 @@
 		"prepack": "run build:ifNeeded",
 		"test:type": "run build:grammar && run tsc",
 		"test": "run build:grammar && mocha \"./test/**/*.test.{js,ts}\"",
+		"bench": "tsup --entry.bench ./bench/index.ts && node ./dist/bench.js",
 		"clean": "rimraf dist node_modules source/grammar.js .deps.json"
 	},
 	"peerDependencies": {