--- a/packages/core/tsconfig.json
+++ b/packages/core/tsconfig.json
@@ -10,5 +10,5 @@
 		"files": true
 	},
 	"exclude": ["./node_modules/", "./dist/"],
-	"include": ["source/**/*.ts"]
+	"include": ["source/**/*.ts", "bench/**/*.ts"]
 }