… examples/nextjs/package.json
… yarn.lock