# Micro optimisation :racehorse: 
Sur la base de règle mon-entreprise, on passe de 1000ms à 730ms pour le parsing des règles soit **37% d'amélioration**

### Comment ?
Quelques micro optimisation avec des for in

#### JS benchmark :
- [for in vs fromEntries(entries().map())](https://jsbenchmark.com/#eyJjYXNlcyI6W3siaWQiOiJTUHd1MzV1cUkyemYzSHhjLTVBeF8iLCJjb2RlIjoiY29uc3QgbyA9IHt9XG5mb3IoY29uc3Qga2V5IGluIERBVEEpIHtcbiAgb1trZXldID0gREFUQVtrZXldLnNyYy50b1N0cmluZygpXG59IiwiZGVwZW5kZW5jaWVzIjpbXX0seyJpZCI6ImFOYzkwdTlDeGpjTnFvVjVCUDVqRCIsImNvZGUiOiJPYmplY3QuZW50cmllcyhEQVRBKS5yZWR1Y2UoKGFjYywgW2ssIHZdKSA9PiB7XG4gIGFjY1trXSA9IHYuc3JjLnRvU3RyaW5nKCk7XG4gIHJldHVybiBhY2M7XG59LCB7fSk7IiwiZGVwZW5kZW5jaWVzIjpbXX0seyJpZCI6IlA3QmVkSzlLclkzd2gwdV8ySnEzYiIsImNvZGUiOiJjb25zdCBvID0ge31cbk9iamVjdC5lbnRyaWVzKERBVEEpLm1hcCgoW2ssIHZdKSA9PiB7IG9ba10gPSB2LnNyYy50b1N0cmluZygpIH0pIiwiZGVwZW5kZW5jaWVzIjpbXX0seyJpZCI6IlJCYWV4TzVUZWNKeXVkZ0RzdUFjbiIsImNvZGUiOiJPYmplY3QuZnJvbUVudHJpZXMoXG4gIE9iamVjdC5lbnRyaWVzKERBVEEpLm1hcCgoZCkgPT4gW2RbMF0sIGRbMV0uc3JjLnRvU3RyaW5nKCldKSxcbilcbiIsImRlcGVuZGVuY2llcyI6W119XSwiY29uZmlnIjp7Im5hbWUiOiJmb3IgaW4gdnMgZnJvbUVudHJpZXMoZW50cmllcygpLm1hcCgpKSIsInBhcmFsbGVsIjp0cnVlLCJnbG9iYWxUZXN0Q29uZmlnIjp7ImRlcGVuZGVuY2llcyI6W119LCJkYXRhQ29kZSI6ImNvbnN0IGRhdGEgPSBbLi4uQXJyYXkoMTAwMDApLmtleXMoKV1cbmNvbnN0IG9iaiA9IHt9XG5kYXRhLmZvckVhY2goKGl0ZW0sIGkpID0+IHtcbiAgY29uc3Qga2V5ID0gJ2tleScgKyBpO1xuICBvYmpba2V5XSA9IHtcbiAgICBzcmM6IE1hdGgucmFuZG9tKCkudG9TdHJpbmcoMTYpLnNsaWNlKDIsIDgpXG4gIH1cbn0pXG5yZXR1cm4gb2JqIn19)
- [for in vs spread operator](https://jsbenchmark.com/#eyJjYXNlcyI6W3siaWQiOiJhTmM5MHU5Q3hqY05xb1Y1QlA1akQiLCJjb2RlIjoiY29uc3Qgb2JqID0ge31cbmZvciAoY29uc3Qga2V5IGluIERBVEEpIHtcblx0b2JqW2tleV0gPSBEQVRBW2tleV1cbn1cbm9iai50ZXN0ID0gNDIiLCJkZXBlbmRlbmNpZXMiOltdfSx7ImlkIjoia0xUd2lOcFNLQWI0anJ6YXl4U19oIiwiY29kZSI6ImNvbnN0IG9iaiA9IHsuLi5EQVRBLH1cbm9iai50ZXN0ID0gNDIiLCJkZXBlbmRlbmNpZXMiOltdfSx7ImlkIjoiX0NBWEZIMHJpU18ydjhvZmJwcjk4IiwiY29kZSI6ImNvbnN0IG9iaiA9IHsuLi5EQVRBLCB0ZXN0OjQyfSIsImRlcGVuZGVuY2llcyI6W119LHsiaWQiOiJMckJfdGpzT1ZQUlIteVE5eE5JSDEiLCJjb2RlIjoiY29uc3Qgb2JqID0gT2JqZWN0LmFzc2lnbih7fSwgREFUQSwge3Rlc3Q6NDJ9KSIsImRlcGVuZGVuY2llcyI6W119XSwiY29uZmlnIjp7Im5hbWUiOiJmb3IgaW4gdnMgc3ByZWFkIG9wZXJhdG9yIiwicGFyYWxsZWwiOnRydWUsImdsb2JhbFRlc3RDb25maWciOnsiZGVwZW5kZW5jaWVzIjpbXX0sImRhdGFDb2RlIjoiY29uc3QgZGF0YSA9IFsuLi5BcnJheSgxMDAwMCkua2V5cygpXVxuY29uc3Qgb2JqID0ge31cbmRhdGEuZm9yRWFjaCgoaXRlbSwgaSkgPT4ge1xuICBjb25zdCBrZXkgPSAna2V5JyArIGk7XG4gIG9ialtrZXldID0ge1xuICAgIHNyYzogTWF0aC5yYW5kb20oKS50b1N0cmluZygxNikuc2xpY2UoMiwgOClcbiAgfVxufSlcbnJldHVybiBvYmoifX0=)


## Résultats 
### Avant
```
benchmark          time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------- -----------------------------
• Parsing initial des règles
----------------------------------------------------- -----------------------------
Modele-social     1.05 s/iter        (1.01 s … 1.1 s)    1.07 s     1.1 s     1.1 s
```

### Après
```
benchmark          time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------- -----------------------------
• Parsing initial des règles
----------------------------------------------------- -----------------------------
Modele-social  729.26 ms/iter  (702.2 ms … 787.81 ms) 744.23 ms 787.81 ms 787.81 ms
```